Create Prims algorithmn
In Prim's algorithm, we divide the graph in two sets of vertices. One set of the explored vertices which have already formed a MST (Set 1) and another set of unexplored vertices which will eventually join the first set to complete "spanning" (Set 2). At each instant, we select a minimum weighted edge in the cut joining the two disjoint sets. If there is no directed edge from explored nodes of the MST to the remaining unexplored nodes, the algorithm gets stuck even though there are edges from unexplored nodes to explored nodes in the MST.
rim's fails because it assumes every node is reachable from every node which, though valid for undirected graphs, may not be true for digraphs. 

However for the given graph, Prims algorithm works fine and the result is obtained.